backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.2

# HTTP client for scraping
httpx==0.27.2
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.1
pytz==2024.1
pydantic==2.9.2
pydantic-settings==2.5.2

# File handling
aiofiles==24.1.0
pillow==10.4.0
